backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python-headless
python-multipart
Pillow
